feat: Add buffer days support for same domain prospect scheduling
This commit adds support for buffer days in the `isSameDomainProspectScheduledForTime` function. The buffer days are used to avoid sending emails to prospects who work in the same company within a few days of each other.

Note: Currently it will add default 2 days of buffer. But later we can make this configurable