feat: Add multithreading support for prospect messaging
Note: This commit adds important comments for Multithreading support added in previous commit

This commit introduces multithreading support in the `isSameDomainProspectScheduledForTime` function. Multithreading allows sending messages to multiple prospects from the same company at different times to avoid spamming. The function checks if the prospect's domain is a generic domain like gmail.com or yahoo.com and returns false. If the prospect is scheduled for the same time as another prospect from the same domain or within a specified number of days, it returns true.

Note: The default buffer is set to 2 days, but it can be made configurable in the future.