feat: set another user as "Reply To" for the email message that we send to prospect as part of campaign.
One of our potential customer wanted to have a different email ID for the “Reply To” field in the email that we send to the prospects as part of the campaign.
If the prospect replies, then another email ID will receive the message. So we won’t be able to detect a reply and show accurate stats.